requirements.txt
fastapi
uvicorn
google-generativeai
supabase
pydantic